Revert endpoint to an old version to allow ESP32 projects to run more reliably
The code in send has been revered to as it was as commit 91e7043.  Commit 217257b changed the code to concatenate the header and body using a string, but for large sends this causes a doubling of the memory required.  A number of web based ESP32 projects nor fail.  This has been documented in issue #131